Moved simulation to a separate python package. Useful for debugging among other things..
