+++ empyer/tests/misc/test_kernels.py
@@ -4,7 +4,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from empyer.misc.kernels import s_g_kernel, s_g_kern_toAng,get_wavelength, sg, shape_function,simulate_symmetry
-from empyer.misc.kernels import random_pattern, simulate_pattern
+from empyer.misc.kernels import random_pattern, simulate_pattern, cartesian_to_ellipse
 
 
 class TestConvert(TestCase):
@@ -36,18 +36,3 @@ class TestConvert(TestCase):
         plt.plot([shape_function(1, (i/100)+.00001)for i in range(100)])
         plt.show()
 
-    def test_simulate_symmetry(self):
-        sim = simulate_symmetry(symmetry=6, iterations=10000)
-        plt.plot(sim[0])
-        plt.show()
-        power_sim = np.mean([np.fft.fft(s).real**2 for s in sim], axis=0)
-        plt.plot(power_sim)
-        plt.show()
-
-    def test_simulation(self):
-        random_pattern(4,4)
-
-    def test_simulate_image(self):
-        i = simulate_pattern(4, k=100, num_clusterns=500, probe_size=20, center=[256, 256], angle=0,  lengths=[10, 10])
-        plt.imshow(i)
-        plt.show()
